Create baseballCosts.cpp
A program that calculates how much a Little League baseball team spent the previous year to purchase new baseballs. The program 
should prompt the user to enter the number of baseballs purchased and the cost of each baseball. It should then calculate and
display the total amount spent to purchase the baseballs.